+++ src/test/java/com/naimish/AddressBook/controller/AddressBookControllerIntegrationTest.java
@@ -44,7 +44,7 @@ class AddressBookControllerIntegrationTest {
     void canAddListAndRemoveContactsInAnAddressBook() throws Exception {
         int addressBookId = createAddressBook("Brisbane");
 
-        int contactId = addContact(addressBookId, "Alice", "111111");
+        long contactId = addContact(addressBookId, "Alice", "111111");
 
         mockMvc.perform(get("/api/addressbooks/" + addressBookId + "/contacts"))
                 .andExpect(status().isOk())
@@ -75,7 +75,7 @@ class AddressBookControllerIntegrationTest {
     void deletingContactViaWrongAddressBookReturnsNotFound() throws Exception {
         int bookOne = createAddressBook("Canberra");
         int bookTwo = createAddressBook("Darwin");
-        int contactId = addContact(bookOne, "Bob", "333333");
+        long contactId = addContact(bookOne, "Bob", "333333");
 
         mockMvc.perform(delete("/api/addressbooks/" + bookTwo + "/contacts/" + contactId))
                 .andExpect(status().isNotFound());
@@ -146,7 +146,7 @@ class AddressBookControllerIntegrationTest {
         return objectMapper.readTree(response).get("id").asInt();
     }
 
-    private int addContact(int addressBookId, String name, String phoneNo) throws Exception {
+    private long addContact(int addressBookId, String name, String phoneNo) throws Exception {
         ContactRequest request = new ContactRequest();
         request.setName(name);
         request.setPhoneNo(List.of(phoneNo));
@@ -157,6 +157,6 @@ class AddressBookControllerIntegrationTest {
                 .andExpect(status().isCreated())
                 .andReturn().getResponse().getContentAsString();
 
-        return objectMapper.readTree(response).get("contactId").asInt();
+        return objectMapper.readTree(response).get("contactId").asLong();
     }
 }
